refactor: 공통 타입들 타입 파일로 이동
