+++ client/src/types/index.ts
@@ -15,3 +15,45 @@ export type SelectOptionFilterType =
   | '편의'
   | '디자인'
   | '주행';
+
+export interface MainOptionType {
+  imgUrl: string;
+  name: string;
+}
+
+export interface ExteriorColorType {
+  imgUrl: string;
+  name: string;
+}
+
+export interface InternalColorType {
+  imgUrl: string;
+  name: string;
+}
+
+export interface TrimType {
+  id: number;
+  name: string;
+  backgroundImgUrl: string;
+  imgUrl: string;
+  hashTag: string;
+  description: string;
+  best: boolean;
+  price: number;
+  mainOptions: MainOptionType[];
+  exteriorColors: ExteriorColorType[];
+  interiorColors: InternalColorType[];
+}
+
+export interface ContentsType {
+  name: string;
+  categoryId: number;
+  imgUrl: string;
+}
+export interface BasicOptionType {
+  first: boolean;
+  last: boolean;
+  totalElements: number;
+  totalPages: number;
+  contents: ContentsType[];
+}
